Rename fault-tolerance extension to smallrye-fault-tolerance
