+++ extensions/smallrye-fault-tolerance/runtime/src/main/java/org/jboss/shamrock/smallrye/faulttolerance/runtime/ShamrockFallbackHandlerProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jboss.shamrock.faulttolerance.runtime;
+package org.jboss.shamrock.smallrye.faulttolerance.runtime;
 
 import javax.annotation.Priority;
 import javax.enterprise.context.Dependent;
