+++ integration-tests/main/pom.xml
@@ -94,7 +94,7 @@
         </dependency>
         <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-fault-tolerance-deployment</artifactId>
+            <artifactId>shamrock-smallrye-fault-tolerance-deployment</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
